% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{addScaleFactors,VisiumExperiment-method}
\alias{addScaleFactors,VisiumExperiment-method}
\alias{addScaleFactors}
\title{addScaleFactors}
\usage{
\S4method{addScaleFactors}{VisiumExperiment}(ve, scaleFactors = list())
}
\arguments{
\item{ve}{a VisiumExperiment class object.}

\item{scaleFactors}{a list of 10x Visium scale factors.-}
}
\value{
a VisiumExperiment class object.
}
\description{
sets the scale factors given as input to the VisiumExperiment
class object.
}
\keyword{internal}
