% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{checkVisiumSpatialCoords,VisiumExperiment-method}
\alias{checkVisiumSpatialCoords,VisiumExperiment-method}
\title{checkVisiumSpatialCoords}
\usage{
\S4method{checkVisiumSpatialCoords}{VisiumExperiment}(ve, spatialCoords = DataFrame())
}
\arguments{
\item{ve}{a VisiumExperiment class object.}

\item{spatialCoords}{a DataFrame with visium spatial coordinates}
}
\value{
a VisiumExperiment class object.
}
\description{
checks if the VisiumExperiment class object has all the required
fields for the integrity of the spatial coordinates and sets them 
to the object.
}
\examples{
TBD
}
