% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\docType{methods}
\name{SPE-combine}
\alias{SPE-combine}
\alias{cbind,SingleCellExperiment-method}
\title{Combining SpatialExperiment objects}
\description{
An overview of methods to combine multiple \linkS4class{SpatialExperiment} 
objects by column.
This method is useful for ensuring that all data fields remain 
synchronized multiple SpatialExperiment objects are combined together, 
such as in case of multiple samples.
}
\section{Combining}{

The \code{...} argument is thought to contain one or more 
\linkS4class{SpatialExperiment} objects.
\describe{
\item{\code{cbind(..., deparse.level=1)}:}{
Returns a SpatialExperiment where
all objects in \code{...} are combined column-wise,
i.e., columns in successive objects are appended to the first object.

Each \code{SpatialExperiment} object in \code{...} must have the same 
\code{colData} (with same \code{spatialCoordinates}).
In case of equal sample_id the method will proceed them providing new 
unique ones.
Additionally the \code{imgData} across all the object are \code{rbind}ed. 

Refer to \code{?"\link{cbind,SingleCellExperiment-method}"} for details on 
how metadata and others inherited attributes are combined in the output 
object.
Refer to \code{?\link[base]{cbind}} for the interpretation of \code{deparse.level}.
}
}
}

\examples{
example(SpatialExperiment, echo=FALSE) # using the class example

# Combining:
cbind(spe, spe)

}
\author{
Dario Righelli
}
