% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.sce_to_se}
\alias{.sce_to_se}
\title{coerce}
\usage{
.sce_to_se(sce, spatialCoords = data.frame())
}
\arguments{
\item{sce}{a SingleCellExperiment object instance}
}
\description{
Converts a SingleCellExperiment into a SpatialExperiment
}
