% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{.se_show}
\alias{.se_show}
\alias{show}
\title{show}
\usage{
.se_show(object)
}
\arguments{
\item{object}{a SpatialExperiment object instance}
}
\description{
a method for showing the SpatialExperiment
}
\examples{
example(SpatialExperiment, echo=FALSE) # using class example
show(se)
}
