% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{.ve_show}
\alias{.ve_show}
\alias{show}
\title{show}
\usage{
.ve_show(object)
}
\arguments{
\item{object}{a VisiumExperiment object instance}
}
\description{
a method for showing the VisiumExperiment
}
\examples{
example(VisiumExperiment, echo=FALSE) #using class example
show(ve)
}
