% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{getCellID,SpatialExperiment-method}
\alias{getCellID,SpatialExperiment-method}
\alias{getCellID}
\title{getCellID}
\usage{
\S4method{getCellID}{SpatialExperiment}(x)
}
\arguments{
\item{x}{a SpatialExperiment object instance}
}
\value{
the cellID
}
\description{
returns the column name where the cell identifiers are be stored
}
\examples{
example(SpatialExperiment)
getCellID(se)
}
