% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{imagePaths<-,VisiumExperiment-method}
\alias{imagePaths<-,VisiumExperiment-method}
\alias{imagePaths<-}
\title{imagePaths-setter}
\usage{
\S4method{imagePaths}{VisiumExperiment}(x) <- value
}
\arguments{
\item{x}{a VisiumExperiment class object}

\item{value}{a list within the paths of the images of a 10x Visium experiment}
}
\value{
none
}
\description{
sets the list of image paths for the VisiumExperiment class 
object.
}
\examples{
example(VisiumExperiment)
imagePaths <- list.files(system.file(file.path("extdata", "10x_visium",
                                   "images"), 
                         package="SpatialExperiment"), full.names=TRUE)
imagePaths(ve) <- imagePaths
}
