% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{imagePaths,VisiumExperiment-method}
\alias{imagePaths,VisiumExperiment-method}
\alias{imagePaths}
\title{imagePaths-getter}
\usage{
\S4method{imagePaths}{VisiumExperiment}(x)
}
\arguments{
\item{x}{a VisiumExperiment class object}
}
\value{
a list of paths of 10x Visium images
}
\description{
getter for the list of imagePaths stored into the 
VisiumExperiment class object.
}
\examples{
example(VisiumExperiment)
imagePaths <- list.files(system.file(file.path("extdata", "10x_visium",
                                   "images"), 
                         package="SpatialExperiment"), full.names=TRUE)
imagePaths(ve)
}
