% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgData-methods.R
\name{imgData-methods}
\alias{imgData-methods}
\alias{loadImg}
\alias{unloadImg}
\alias{addImg}
\alias{removeImg}
\alias{imgPath,SpatialExperiment-method}
\alias{imgUrl,SpatialExperiment-method}
\alias{loadImg,SpatialExperiment-method}
\alias{unloadImg,SpatialExperiment-method}
\alias{addImg,SpatialExperiment-method}
\alias{removeImg,SpatialExperiment-method}
\title{Methods for handling image-related data}
\usage{
\S4method{imgGrob}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{imgPath}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{imgUrl}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{loadImg}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{unloadImg}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{addImg}{SpatialExperiment}(x, imageSource, scaleFactor, sample_id, image_id, load = TRUE)

\S4method{removeImg}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)
}
\arguments{
\item{x}{a \code{\link{SpatialExperiment}}}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying 
sample/image identifier(s); here, \code{TRUE} is equivalent to all 
samples/images and \code{NULL} specifies the first available entry (see details)}

\item{image_id}{see \code{sample_id}}

\item{imageSource}{a character string specifying an image file name 
(.png, .jpg or .tif) or URL to source the image from}

\item{scaleFactor}{single numeric scale factor used to rescale spatial 
coordinates according to the image's resolution}

\item{load}{logical; should the image(s) be loaded as a \code{grob}?
if FALSE, will store the path/URL instead}
}
\value{
\code{add/removeImg()} return a \code{\link{SpatialExperiment}} 
with modified \code{imgData}; specifically, they create/remove 
an image entry (row) in the \code{imgData} \code{DataFrame}.

\code{imgGrob/Path/Url()} access relevant data in the \code{SpatialImage}(s)
stored inside the \code{imgData}'s \code{data} field. 
Depending on whether or not multiple entries are excesses,
a character string or vector is returned by \code{imgPath/Url()},
and a single or list of \code{grob}(s) is returned by \code{imgGrob()}.
  
#' @examples
data(ve)
(df <- imgData(ve))

# 'SpatialImage' accessors
(si <- df$data[[1]])
imgGrob(si)
imgPath(si)
imgUrl(si)

# unload all images
ve <- unloadImg(ve,
  sample_id = TRUE,
  image_id = TRUE)
imgData(ve)$data

# reload all images
ve <- loadImg(ve,
  sample_id = TRUE,
  image_id = TRUE)
imgData(ve)$data

# remove an image
ve <- removeImg(ve,
  sample_id = "section1",
  image_id = "lowres")
imgData(ve)

# add an image
url <- "https://i.redd.it/3pw5uah7xo041.jpg"
ve <- addImg(ve,
  sample_id = "section1",
  image_id = "pomeranian",
  imageSource = url,
  scaleFactor = NA_real_,
  load = TRUE)

grb <- imgGrob(ve,
  sample_id = "section1",
  image_id = "pomeranian")
grid::grid.draw(grb)
}
\description{
The set of functions described below is designed to handle 
the image-related data stored inside a \code{SpatialExperiment}'s 
\code{imgData} \code{int_metadata} field. These include 
\itemize{
\item{\code{imgGrob/Path/Url()}
  to access the \code{grob}, path and URL 
  associated with an image or set of images}
\item{\code{loadImg()} 
  to load an image from a path or URL as a \code{grob}}
\item{\code{unloadImg()} 
  to unload an image, i.e. to drop the \code{grob}
  while retaining the image's source path and/or URL}
\item{\code{add/removeImg()} 
  to add/remove an image entry to/from 
  the \code{imgData} \code{DataFrame}}
}
}
\author{
Helena L. Crowell
}
