% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{isInTissue,SpatialExperiment-method}
\alias{isInTissue,SpatialExperiment-method}
\title{isInTissue}
\usage{
\S4method{isInTissue}{SpatialExperiment}(x, sample_id = TRUE)
}
\arguments{
\item{x}{a VisiumExperiment class object.}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying sample 
identifier(s); here, \code{TRUE} is equivalent to all samples 
and \code{NULL} specifies the first available entry (see details)}
}
\value{
a TRUE/FALSE mask.
}
\description{
returns a mask of TRUE/FALSE Barcodes spots, indicating which
ones are in tissue and which ones are not.
}
\examples{
data(ve)
isInTissue(ve)
sum(isInTissue(ve))
ve[isInTissue(ve),]
}
