% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10xVisium.R
\name{read10xVisium}
\alias{read10xVisium}
\title{Load data from a 10X Visium experiment}
\usage{
read10xVisium(
  samples = "",
  sample_id = paste0("sample", seq_along(samples)),
  type = c("HDF5", "sparse"),
  images = c("fullres", "lowres", "hires", "detected", "aligned"),
  load = TRUE
)
}
\arguments{
\item{samples}{a character vector specifying one or more directories, 
each corresponding to a 10X Visium sample (see details);
if provided, names will be used as sample identifiers}

\item{sample_id}{character string specifying unique sample identifiers,
one for each directory specified via \code{samples}; 
ignored if \code{!is.null(names(samples))}}

\item{type}{character string specifying 
the type of format to read count data from
(see \code{\link{read10xCounts}})}

\item{images}{character vector specifying which images to include.}

\item{load}{logical; should the image(s) be loaded into memory
as a \code{grob}? If FALSE, will store the path/URL instead.}
}
\value{
a \code{\link{SpatialExperiment}}
}
\description{
Creates a \code{\link{SpatialExperiment}} from the CellRanger 
output directories for 10X Visium spatial gene expression data.
}
\details{
The constructor assumes data from each sample are located 
in a single output directory as returned by Space Range, 
thus having the following file organization:

sample \cr
|—outs \cr
· · |—raw/filtered_feature_bc_matrix.h5 \cr
· · |—raw/filtered_feature_bc_matrix    \cr
· · · · |—barcodes.tsv \cr
· · · · |—features.tsv \cr
· · · · |—matrix.mtx   \cr
· · |—spatial \cr
· · · · |—scalefactors_json.json    \cr
· · · · |—tissue_lowres_image.png   \cr
· · · · |—tissue_positions_list.csv \cr
}
\examples{
dir <- system.file(
  file.path("extdata", "10xVisium"), 
  package = "SpatialExperiment")
  
sample_ids <- c("section1", "section2")
samples <- file.path(dir, sample_ids)
  
list.files(samples[1])
list.files(file.path(samples[1], "spatial"))
file.path(samples[1], "raw_feature_bc_matrix")

(ve <- read10xVisium(samples, sample_ids, type="HDF5",
  images = c("lowres"), load = FALSE))

# tabulate number of spots mapped to tissue
table(
  in_tissue = inTissue(ve), 
  sample_id = ve$sample_id )

# view available images
imgData(ve)

}
\author{
Helena L. Crowell
}
