% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImgData.R
\name{readImgData}
\alias{readImgData}
\title{Read images & scale factors}
\usage{
readImgData(
  path = ".",
  sample_id = names(path),
  imageSources = file.path(path, "tissue_lowres_image.png"),
  scaleFactors = file.path(path, "scalefactors_json.json"),
  load = TRUE
)
}
\arguments{
\item{path}{...}

\item{scaleFactors}{...}

\item{load}{logical; should the image(s) be loaded into memory
as a \code{grob}? If FALSE, will store the path/URL instead.}

\item{sampleId}{...}

\item{images}{...}
}
\value{
a \code{\link{DataFrame}}
}
\description{
Read images & scale factors
}
\examples{
dir <- system.file(
  file.path("extdata", "10xVisium", "spatial"),
  package = "SpatialExperiment")
  
# base directory contains 
# - scale factors (scalefactors_json.json)
# - one image (tissue_lowres_image.png)
list.files(dir)

# read in images & scale factors
# as valid 'imgData' 'DFrame'
readImgData(dir, sampleId = "foo")

}
\author{
Helena L. Crowell
}
