% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{scaleFactors<-,VisiumExperiment-method}
\alias{scaleFactors<-,VisiumExperiment-method}
\alias{scaleFactors<-}
\title{scaleFactors-setter}
\usage{
\S4method{scaleFactors}{VisiumExperiment}(x) <- value
}
\arguments{
\item{x}{a VisiumExperiment class object.}

\item{value}{a list of 10x Visium scale factors.}
}
\value{
a VisiumExperiment class object.
}
\description{
sets the scale factors in a VisiumExperiment class object.
}
\examples{

example(VisiumExperiment)
}
