% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{scaleFactors,VisiumExperiment-method}
\alias{scaleFactors,VisiumExperiment-method}
\alias{scaleFactors}
\title{scaleFactors-getter}
\usage{
\S4method{scaleFactors}{VisiumExperiment}(x)
}
\arguments{
\item{x}{a VisiumExperiment class object.}
}
\value{
a DataFrame with the 10x Visium scale factors.
}
\description{
gets the scale factors from a VisiumExperiment class object.
}
\examples{
ve <- readRDS(file=system.file(file.path("extdata", "10x_visium",
                         "ve.RDS"), package="SpatialExperiment"))
scaleFactors(ve)
}
