% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{show,SpatialExperiment-method}
\alias{show,SpatialExperiment-method}
\alias{show}
\title{SpatialExperiment show method}
\usage{
\S4method{show}{SpatialExperiment}(object)
}
\arguments{
\item{object}{a SpatialExperiment object instance}
}
\value{
none
}
\description{
a method for showing the SpatialExperiment
}
\examples{
example(SpatialExperiment, echo=FALSE) # using class example
show(se)
}
