% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{show,VisiumExperiment-method}
\alias{show,VisiumExperiment-method}
\alias{.ve_show}
\title{VisiumExperiment show method}
\usage{
\S4method{show}{VisiumExperiment}(object)
}
\arguments{
\item{object}{a VisiumExperiment object instance}
}
\value{
none
}
\description{
a method for showing the VisiumExperiment
}
\examples{
example(VisiumExperiment, echo=FALSE) #using class example
show(ve)
}
