% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{spatialCoords<-,VisiumExperiment-method}
\alias{spatialCoords<-,VisiumExperiment-method}
\title{spatialCoords-setter}
\usage{
\S4method{spatialCoords}{VisiumExperiment}(x) <- value
}
\arguments{
\item{x}{a VisiumExperiment class object.}

\item{value}{a DataFrame of spatial coordinates}
}
\value{

}
\description{
sets the spatial coordinate to a VisiumExperiment class object.
}
\examples{
TBD
}
