% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{spatialCoords<-}
\alias{spatialCoords<-}
\title{#' @export
setGeneric("scaleFactors<-", function(x, value)
    standardGeneric("scaleFactors<-"))}
\usage{
spatialCoords(x, ...) <- value
}
\description{
#' @export
setGeneric(name="spatialCoords", def=function(se, ...)
    standardGeneric("spatialCoords"))
}
\details{
#' @export
setGeneric(name="spatialCoordsMtx", def=function(se, ...)
    standardGeneric("spatialCoordsMtx"))
}
