% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{spatialCoords,VisiumExperiment-method}
\alias{spatialCoords,VisiumExperiment-method}
\title{spatialCoords-getter}
\usage{
\S4method{spatialCoords}{VisiumExperiment}(x)
}
\arguments{
\item{x}{a  VisiumExperiment class object.}
}
\value{
a DataFrame with the spatial coordinates.
}
\description{
gets the spatial coordinate from a VisiumExperiment class object
}
\examples{
TBD
}
