% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{spatialCoords}
\alias{spatialCoords}
\title{#' @export
setGeneric("scaleFactors<-", function(x, value)
    standardGeneric("scaleFactors<-"))}
\usage{
spatialCoords(se, sample_id)
}
\description{
#' @export
setGeneric("scaleFactors<-", function(x, value)
    standardGeneric("scaleFactors<-"))
}
