% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialCoordsMtx,SpatialExperiment-method}
\alias{spatialCoordsMtx,SpatialExperiment-method}
\title{spatialCoordsMtx}
\usage{
\S4method{spatialCoordsMtx}{SpatialExperiment}(se, sample_id = TRUE)
}
\arguments{
\item{se}{A SpatialExperiment class object.}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying sample 
identifier(s); here, \code{TRUE} is equivalent to all samples 
and \code{NULL} specifies the first available entry (see details)}
}
\value{
a matrix object within the spatial coordinates.
}
\description{
a getter method which returns the spatial coordinates previously
stored in a SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
spatialCoordsMtx(se)
}
