% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{spatialCoordsNames,SpatialExperiment-method}
\alias{spatialCoordsNames,SpatialExperiment-method}
\alias{spatialCoordsNames}
\title{spatialCoordsNames-getter}
\usage{
\S4method{spatialCoordsNames}{SpatialExperiment}(x)
}
\arguments{
\item{x}{a SpatialExperiment class object.}
}
\value{
a vector with the colnames of the spatial coordinates.
}
\description{
getter method for the spatial coordinates names in a 
SpatialExperiment class object.
}
\examples{
spatialDataFile <- system.file(file.path("extdata", "seqFISH",
    "seqFISH.RData"), package="SpatialExperiment")
load(spatialDataFile)
se <- SpatialExperiment(rowData=rownames(fishFeaturesCounts),
    colData=fishCellLabels,
    assays=SimpleList(counts=as.matrix(fishFeaturesCounts)),
    spatialCoords=fishCoordinates)
spatialCoordsNames(se)
}
