% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{spatialCoordsNames}
\alias{spatialCoordsNames}
\title{#' @export
setGeneric("isInTissue", function(x) standardGeneric("isInTissue"))}
\usage{
spatialCoordsNames(x)
}
\description{
#' @export
setGeneric("isInTissue", function(x) standardGeneric("isInTissue"))
}
