% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialData.R
\name{.get_row_idx}
\alias{.get_row_idx}
\alias{spatialData}
\alias{spatialCoords}
\alias{spatialDataNames}
\alias{spatialCoords,SpatialExperiment-method}
\alias{spatialDataNames,SpatialExperiment-method}
\title{Methods for handling spatial-related data}
\usage{
.get_row_idx(x, sample_id)

\S4method{spatialCoords}{SpatialExperiment}(x, cd_keep = NULL, sample_id = TRUE, as.df = FALSE)

\S4method{spatialDataNames}{SpatialExperiment}(x)
}
\arguments{
\item{x}{a \code{\link{SpatialExperiment}}.}

\item{sample_id}{specifies which sample(s) to keep; valid values are 
TRUE (default) for all samples, NULL for the first entry available, 
or a character vector with elements in \code{unique(x$sample_id)}.}

\item{cd_keep}{character vector specifying 
additional \code{colData} columns to retain.}

\item{as.df}{logical; should the retrieved data be converted 
to a \code{data.frame}? By default, they are returned as a matrix.}
}
\value{
\describe{
\item{spatialData/Coords}{
  a \code{DataFrame/matrix} of all spatial-related 
  data or spatial coordinates, respectively.
  When \code{as.df = TRUE}, data are returned as a \code{data.frame};
  when \code{cd_keep} is specified, these are included as well.}
\item{spatialDataNames}{
  a character vector of spatial-related data names. 
  E.g. for spot-based experiments, these include
  xyz-coordinate and array position names.}
}
}
\description{
The set of functions described below is designed to handle spatial-related 
data stored inside a \code{SpatialExperiment}'s \code{colData}, including:
\itemize{
\item{\code{spatialData/Coords()} 
  to retrieve a table of all spatially related data or 
  spatial coordinates, respectively, and (optionally) other
  observation metadata (\code{colData}) variables of interest.}
\item{\code{spatialDataNames()}
  to retrieve a character vector of spatial coordinate names}
}
}
\examples{
example(SpatialExperiment)
spatialDataNames(se)
head(spatialData(se))
head(spatialCoords(se))
}
