% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialData<-,SpatialExperiment-method}
\alias{spatialData<-,SpatialExperiment-method}
\title{spatialData<-}
\usage{
\S4method{spatialData}{SpatialExperiment}(x, sample_id = TRUE) <- value
}
\arguments{
\item{x}{a SpatialExperiment class object}

\item{value}{a DataFrame with the new spatial data to set (see details).}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying sample 
identifier(s); here, \code{TRUE} is equivalent to all samples 
and \code{NULL} specifies the first available entry (see details)}
}
\value{
none
}
\description{
a setter method which sets/replaces the spatial data in a
SpatialExperiment class object.
They are always stored as \code{x_coord}, \code{y_coord} and when found 
\code{in_tissue}, \code{array_row} and \code{array_col} (see details).
}
\details{
The method automatically recognizes any kind of colnames with
\code{x}, \code{y}, \code{z}, \code{x_coord}, \code{y_coord}, \code{z_coord} and, 
additionally, for the 10x Visium, it recognizes the 
\code{pxl_row_in_fullres} and \code{pxl_col_in_fullres} storing them respectively as 
\code{y_coord} and \code{x_coord}.
While \code{in_tissue} \code{array_row} \code{array_col} are stored as they are.
NB \code{in_tissue} have to be in logical form.
}
\examples{
example(SpatialExperiment)
fakeCoords <- cbind(spatialData(se)$x_coord, spatialData(se)$x_coord)
colnames(fakeCoords) <- c("x", "y")
spatialData(se) <- fakeCoords
spatialData(se)
oneCoord <- cbind(spatialData(se)$y_coord)
colnames(oneCoord) <- c("pxl_row_in_fullres") # assigns it to the y_coord
spatialData(se) <- oneCoord
spatialData(se)
}
