% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialData,SpatialExperiment-method}
\alias{spatialData,SpatialExperiment-method}
\title{spatialData}
\usage{
\S4method{spatialData}{SpatialExperiment}(se, sample_id = TRUE)
}
\arguments{
\item{se}{A SpatialExperiment class object.}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying sample 
identifier(s); here, \code{TRUE} is equivalent to all samples 
and \code{NULL} specifies the first available entry (see details)}
}
\value{
a DataFrame within the spatial coordinates.
}
\description{
a getter method which returns the spatial coordinates 
data structure previously stored in a SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
spatialData(se)
}
