% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialDataNames,SpatialExperiment-method}
\alias{spatialDataNames,SpatialExperiment-method}
\title{spatialDataNames}
\usage{
\S4method{spatialDataNames}{SpatialExperiment}(x)
}
\arguments{
\item{x}{a SpatialExperiment class object.}
}
\value{
a vector with the colnames of the spatial coordinates.
}
\description{
getter method for the spatial coordinates names in a
SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
spatialDataNames(se)
}
