% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-colData.R
\name{SpatialExperiment-colData}
\alias{SpatialExperiment-colData}
\alias{colData}
\alias{colData<-}
\alias{colData<-,SpatialExperiment,DataFrame-method}
\alias{colData<-,SpatialExperiment,NULL-method}
\title{SpatialExperiment colData}
\usage{
\S4method{colData}{SpatialExperiment,DataFrame}(x) <- value

\S4method{colData}{SpatialExperiment,`NULL`}(x) <- value
}
\arguments{
\item{x}{a \code{\link{SpatialExperiment}}}

\item{value}{a \code{\link[S4Vectors]{DataFrame}}}
}
\value{
a \code{\link{SpatialExperiment}} object with updated \code{colData}
}
\description{
The \code{\link{SpatialExperiment}} class provides a modified \code{colData} setter, which ensures 
that the \code{SpatialExperiment} object remains valid.
}
\details{
The \code{colData} setter expects a \code{\link{DataFrame}} with a
\code{sample_id} column reflecting the already existing \code{sample_id}(s)
present in the \code{SpatialExperiment} object. An additional check is
performed against the \code{\link{imgData}} data structure. If a \code{NULL}
value is provided, the \code{colData} is dropped.
}
\examples{
example(read10xVisium)
colData(spe) <- NULL
}
