% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-combine.R
\docType{methods}
\name{SpatialExperiment-combine}
\alias{SpatialExperiment-combine}
\alias{cbind,SingleCellExperiment-method}
\alias{cbind,SpatialExperiment-method}
\title{Combining SpatialExperiment objects}
\usage{
\S4method{cbind}{SpatialExperiment}(..., deparse.level = 1)
}
\arguments{
\item{...}{a list of SpatialExperiment objects}

\item{deparse.level}{Refer to \code{?\link[base]{rbind}}}
}
\value{
a combined SpatialExperiment object
}
\description{
An overview of methods to combine multiple \linkS4class{SpatialExperiment} 
objects by column.
This method is useful for ensuring that all data fields remain 
synchronized multiple SpatialExperiment objects are combined together, 
such as in case of multiple samples.
}
\section{Combining}{

The \code{...} argument is thought to contain one or more 
\linkS4class{SpatialExperiment} objects.

\describe{
\item{\code{cbind(..., deparse.level=1)}:}{ 
Returns a SpatialExperiment where all objects in \code{...}
are combined column-wise, i.e., columns in successive objects 
are appended to the first object.

Each \code{SpatialExperiment} object in \code{...} must have the same 
\code{colData} (with same \code{spatialCoordinates}).
In case of equal sample_id the method will proceed them providing new 
unique ones. 
Additionally the \code{imgData} across all the object are \code{rbind}ed. 

Refer to \code{?"\link{cbind,SingleCellExperiment-method}"} 
for details on how metadata and others inherited attributes 
are combined in the output object.

Refer to \code{?\link[base]{cbind}} 
for the interpretation of \code{deparse.level}.
}
}
}

\examples{
example(read10xVisium, echo = FALSE)

# merging with duplicated 'sample_id's
# will automatically make them unique
spe1 <- spe2 <- spe
spe3 <- cbind(spe1, spe2)
unique(spe3$sample_id)

# make sample identifiers unique
spe1 <- spe2 <- spe
spe1$sample_id <- paste(spe1$sample_id, "sample1", sep = ".")
spe2$sample_id <- paste(spe2$sample_id, "sample2", sep = ".")

# combine into single object
spe <- cbind(spe1, spe2)

# view joint 'imgData'
imgData(spe)

# tabulate number of spots mapped to tissue
table(
  in_tissue = spe$in_tissue, 
  sample_id = spe$sample_id)
}
\author{
Dario Righelli
}
