% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{SpatialExperiment-misc}
\alias{SpatialExperiment-misc}
\alias{.spe_show}
\alias{show,SpatialExperiment-method}
\title{Miscellaneous \code{SpatialExperiment} methods}
\usage{
.spe_show(object)

\S4method{show}{SpatialExperiment}(object)
}
\arguments{
\item{object}{a \code{SpatialExperiment}.}
}
\value{
none.
}
\description{
Miscellaneous methods for the \code{\link{SpatialExperiment}} class
and its descendants that do not fit in any other documentation category
such as, for example, show methods.
}
\author{
Dario Righelli & Helena L. Crowell
}
