% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{SpatialExperiment-subset}
\alias{SpatialExperiment-subset}
\alias{[,SpatialExperiment,ANY,ANY,ANY-method}
\title{SpatialExperiment subset}
\arguments{
\item{x}{a SpatialExperiment object}

\item{i}{the row indexes to subset}

\item{j}{the col indexes to subset}
}
\value{
a SpatialExperiment class object
}
\description{
The subset method ensures that all the subsetted rows/cols 
are updated across the colData and spatialData objects
}
\section{subset}{

\describe{
\item{\code{[}:}{ subsetting method}
}
}

\examples{
example(SpatialExperiment)
idx <- sample(ncol(spe), 10)
sub <- spe[, idx]
dim(sub)
spatialData(sub, 
  spatialCoords = TRUE, 
  colData = TRUE)
}
