% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SpatialExperiment-class}
\alias{SpatialExperiment-class}
\alias{SpatialExperiment}
\alias{coerce,SingleCellExperiment,SpatialExperiment-method}
\title{The SpatialExperiment class}
\usage{
SpatialExperiment(..., spatialCoords = data.frame())
}
\arguments{
\item{...}{arguments to be passed to the \code{\link{SingleCellExperiment}} 
constructor to fill the slots of the base class.}

\item{spatialCoords}{the 10x Visium spatial coordinates}
}
\description{
The SpatialExperiment class is designed to represent 10x Visium spatial 
Gene Expression data.
It inherits from the \linkS4class{SingleCellExperiment} class and is used in 
the same manner.
In addition, the class supports the integration with 10x Visium spatial 
coordinates and its scale factors.
}
\section{Slots}{

\describe{
\item{\code{int_spcIdx}}{integer.}
}}

\examples{
fishCoordFile <- system.file(file.path("extdata", "seqFISH",
                               "fcortex.coordinates.txt"), 
                               package="SpatialExperiment")
fishCoordinates <- read.table(fishCoordFile, header=FALSE, sep=" ")
colnames(fishCoordinates) <- c("ID", "Irrelevant", "x", "y")

fishCellLabsFile <- system.file(file.path("extdata", "seqFISH", 
                                         "seqfish_cell_labels.tsv"),
                            package="SpatialExperiment")
fishCellLabels <- read.table(file=fishCellLabsFile, header=FALSE, sep="\t")
colnames(fishCellLabels) <- c("ID", "cluster", "class", "classID", 
                              "Irrelevant", "Prob")
fishFeatCountsFile <- system.file(file.path("extdata", "seqFISH",
                             "seqfish_normalized_cortex_b2_testing.txt"), 
                             package="SpatialExperiment")
fishFeaturesCounts <- read.table(file=fishFeatCountsFile, 
                                 header=FALSE, sep="\t", row.names=1)
 se <- SpatialExperiment(rowData=rownames(fishFeaturesCounts),
                     colData=fishCellLabels,
                     assays=SimpleList(counts=as.matrix(fishFeaturesCounts)),
                     spatialCoords=fishCoordinates) 
}
\author{
Dario Righelli
}
