% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialImage-methods.R
\name{SpatialImage-methods}
\alias{SpatialImage-methods}
\alias{imgGrob}
\alias{imgPath}
\alias{imgUrl}
\alias{imgGrob<-}
\alias{imgPath<-}
\alias{imgUrl<-}
\alias{imgPath,SpatialImage-method}
\alias{imgUrl,SpatialImage-method}
\alias{imgGrob<-,SpatialImage,rastergrob-method}
\alias{imgPath<-,SpatialImage,character-method}
\alias{imgUrl<-,SpatialImage,character-method}
\alias{loadImg,SpatialImage-method}
\alias{unloadImg,SpatialImage-method}
\title{SpatialImage methods}
\usage{
\S4method{imgGrob}{SpatialImage}(x)

\S4method{imgPath}{SpatialImage}(x)

\S4method{imgUrl}{SpatialImage}(x)

\S4method{imgGrob}{SpatialImage,rastergrob}(x) <- value

\S4method{imgPath}{SpatialImage,character}(x) <- value

\S4method{imgUrl}{SpatialImage,character}(x) <- value

\S4method{loadImg}{SpatialImage}(x)

\S4method{unloadImg}{SpatialImage}(x)
}
\arguments{
\item{x}{a \code{\link{SpatialImage}}}

\item{value}{a \code{grob} for \code{imgGrob}, 
a character string for \code{imgPath/Url}}
}
\description{
Getters, setters and additional methods 
for the \code{\link{SpatialImage}} class.
}
\details{
In the following examples, \code{x} is a \code{\link{SpatialImage}}.
\describe{
\item{getters & setters:}{
\itemize{
\item{\code{imgGrob(x), imgGrob(x) <- value} \cr gets/sets the image's \code{grob}}
\item{\code{imgPath(x), imgPath(x) <- value} \cr gets/sets the image's file path}
\item{\code{imgUrl(x), imgUrl(x) <- value} \cr gets/sets the image's source URL}
}  
}
\item{image methods:}{
\itemize{
\item{\code{loadImg(x)} loads an image as a \code{grob} from its path or URL}
\item{\code{unloadImg(x)} unloads an image by dropping the \code{grob}}
} 
}
}
}
\examples{
# construct 'SpatialImage'
dir <- file.path("extdata", "10xVisium", "section1", "spatial")
fnm <- file.path(dir, "tissue_lowres_image.png")
imgPath <- system.file(fnm, package = "SpatialExperiment")
(img <- SpatialImage(path = imgPath))

# load image as a 'grob'
img <- loadImg(img)

# accessors
imgGrob(img)
imgPath(img)
imgUrl(img)

# draw the image
grb <- imgGrob(img)
grid::grid.draw(grb)

# unload it, i.e. the 'grob'
img <- unloadImg(img)
imgGrob(img)
  
}
\author{
Helena L. Crowell
}
