% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve.R
\docType{data}
\name{ve}
\alias{ve}
\title{A SpatialExperiment object with two samples of 10x Visium data loaded}
\format{
A SpatialExperiment object of two 10x Visium Samples
}
\usage{
ve
}
\description{
A SpatialExperiment object with two samples of 10x Visium data loaded
}
\keyword{datasets}
