% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentC.R
\docType{class}
\name{VisiumExperiment-class}
\alias{VisiumExperiment-class}
\alias{VisiumExperiment}
\title{The VisiumExperiment class}
\usage{
VisiumExperiment(..., spatialCoords = data.frame(), scaleFactors = list())
}
\arguments{
\item{...}{arguments to be passed to the \code{\link{SingleCellExperiment}} 
constructor to fill the slots of the base class.}

\item{spatialCoords}{the 10x Visium spatial coordinates}

\item{scaleFactors}{the 10x Visium image scale factors}
}
\description{
The VisiumExperiment class is designed to represent 10x Visium spatial 
Gene Expression data.
It inherits from the \linkS4class{SingleCellExperiment} class and is used in 
the same manner.
In addition, the class supports the integration with 10x Visium spatial 
coordinates and its scale factors.
}
\section{Slots}{

\describe{
\item{\code{scaleFactors}}{list.}

\item{\code{int_spcIdx}}{integer.}
}}

\examples{
TBD
}
\author{
Dario Righelli
}
