% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{checkSpatialCoords,SpatialExperiment-method}
\alias{checkSpatialCoords,SpatialExperiment-method}
\title{checkSpatialCoords}
\usage{
\S4method{checkSpatialCoords}{SpatialExperiment}(se, spatialCoords = DataFrame())
}
\arguments{
\item{SpatialExperiment}{a SpatialExperiment class object}
}
\value{
the SpatialExperiment class object passed as input.
}
\description{
checks if all the spatial parameters have the right fields.
}
\keyword{internal}
