% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{.setCoord}
\alias{.setCoord}
\title{#' @rdname SpatialExperiment-methods
#' @export
setReplaceMethod("scaleFactors", 
                 c("SpatialExperiment", "list"),
                 function(x, sample_id=TRUE, image_id=TRUE, 
                          scaleFactors=as.list(rep(1,1)))
                 {
                     idx <- .get_img_idx(x, sample_id, image_id)
                     imgData(x)$scaleFactor[idx] <- scaleFactors
                 }
)}
\usage{
.setCoord(df, coordName, coordinates)
}
\description{
#' @rdname SpatialExperiment-methods
#' @export
setReplaceMethod("scaleFactors", 
                 c("SpatialExperiment", "list"),
                 function(x, sample_id=TRUE, image_id=TRUE, 
                          scaleFactors=as.list(rep(1,1)))
                 {
                     idx <- .get_img_idx(x, sample_id, image_id)
                     imgData(x)$scaleFactor[idx] <- scaleFactors
                 }
)
}
