% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{getCellID,SpatialExperiment-method}
\alias{getCellID,SpatialExperiment-method}
\title{getCellID}
\usage{
\S4method{getCellID}{SpatialExperiment}(x)
}
\arguments{
\item{SpatialExperiment}{A SpatialExperiment object}
}
\value{
the CellID name
}
\description{
getCellID
}
\keyword{internal}
