% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialImage-methods.R, R/imgData-methods.R
\name{imgPath,SpatialImage-method}
\alias{imgPath,SpatialImage-method}
\alias{imgGrob,SpatialImage-method}
\alias{imgUrl,SpatialImage-method}
\alias{imgPath<-,SpatialImage,character-method}
\alias{imgGrob<-,SpatialImage,rastergrob-method}
\alias{imgUrl<-,SpatialImage,character-method}
\alias{loadImg,SpatialImage-method}
\alias{unloadImg,SpatialImage-method}
\alias{imgData-methods}
\alias{loadImg}
\alias{unloadImg}
\alias{addImg}
\alias{removeImg}
\alias{imgGrob}
\alias{imgPath,ANY}
\alias{imgUrl}
\alias{imgGrob<-}
\alias{imgPath<-,ANY}
\alias{imgUrl<-}
\alias{imgGrob,SpatialExperiment-method}
\alias{imgUrl,SpatialExperiment-method}
\alias{loadImg,SpatialExperiment-method}
\alias{unloadImg,SpatialExperiment-method}
\alias{addImg,SpatialExperiment-method}
\alias{removeImg,SpatialExperiment-method}
\title{Methods for handling image-related data}
\usage{
\S4method{imgPath}{SpatialImage}(x)

\S4method{imgGrob}{SpatialImage}(x)

\S4method{imgUrl}{SpatialImage}(x)

\S4method{imgPath}{SpatialImage,character}(x) <- value

\S4method{imgGrob}{SpatialImage,rastergrob}(x) <- value

\S4method{imgUrl}{SpatialImage,character}(x) <- value

\S4method{loadImg}{SpatialImage}(x)

\S4method{unloadImg}{SpatialImage}(x)

\S4method{imgPath}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{imgGrob}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{imgUrl}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{loadImg}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{unloadImg}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)

\S4method{addImg}{SpatialExperiment}(x, imageSource, scaleFactor, sample_id, image_id, load = TRUE)

\S4method{removeImg}{SpatialExperiment}(x, sample_id = NULL, image_id = NULL)
}
\arguments{
\item{x}{a \code{\link{SpatialExperiment}} or a \code{\link{SpatialImage}}}

\item{value}{a \code{grob} for \code{imgGrob}, 
a character string for \code{imgPath/Url}}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying 
sample/image identifier(s); here, \code{TRUE} is equivalent to all 
samples/images and \code{NULL} specifies the first available entry (see details)}

\item{image_id}{see \code{sample_id}}

\item{imageSource}{a character string specifying an image file name 
(.png, .jpg or .tif) or URL to source the image from}

\item{scaleFactor}{single numeric scale factor used to rescale spatial 
coordinates according to the image's resolution}

\item{load}{logical; should the image(s) be loaded as a \code{grob}?
if FALSE, will store the path/URL instead}
}
\value{
\code{add/removeImg()} return a \code{\link{SpatialExperiment}} 
with modified \code{imgData}; specifically, they create/remove 
an image entry (row) in the \code{imgData} \code{DataFrame}.

\code{imgGrob/Path/Url()} access relevant data in the \code{SpatialImage}(s)
stored inside the \code{imgData}'s \code{data} field. 
Depending on whether or not multiple entries are excesses,
a character string or vector is returned by \code{imgPath/Url()},
and a single or list of \code{grob}(s) is returned by \code{imgGrob()}.
}
\description{
The set of functions described below is designed to handle 
the image-related data stored inside a \code{SpatialExperiment}'s 
\code{imgData} \code{int_metadata} field. These include 
\itemize{
\item{\code{imgGrob/Path/Url()}
  to access the \code{grob}, path and URL 
  associated with an image or set of images}
\item{\code{loadImg()} 
  to load an image from a path or URL as a \code{grob}}
\item{\code{unloadImg()} 
  to unload an image, i.e. to drop the \code{grob}
  while retaining the image's source path and/or URL}
\item{\code{add/removeImg()} 
  to add/remove an image entry to/from 
  the \code{imgData} \code{DataFrame}}
}
\describe{
Getters, setters and additional methods 
for the \code{\link{SpatialImage}} class.
\item{getters & setters:}{
\itemize{
\item{\code{imgGrob(x), imgGrob(x) <- value} \cr gets/sets the image's \code{grob}}
\item{\code{imgPath(x), imgPath(x) <- value} \cr gets/sets the image's file path}
\item{\code{imgUrl(x), imgUrl(x) <- value} \cr gets/sets the image's source URL}
}  
}
\item{image methods:}{
\itemize{
\item{\code{loadImg(x)} loads an image as a \code{grob} from its path or URL}
\item{\code{unloadImg(x)} unloads an image by dropping the \code{grob}}
} 
}
}
}
\examples{
data(ve)
(df <- imgData(ve))

# 'SpatialImage' accessors
(si <- df$data[[1]])
imgGrob(si)
imgPath(si)
imgUrl(si)

# unload all images
ve <- unloadImg(ve,
  sample_id = TRUE,
  image_id = TRUE)
imgData(ve)$data

# reload all images
ve <- loadImg(ve,
  sample_id = TRUE,
  image_id = TRUE)
imgData(ve)$data

# remove an image
ve <- removeImg(ve,
  sample_id = "section1",
  image_id = "lowres")
imgData(ve)

# add an image
url <- "https://i.redd.it/3pw5uah7xo041.jpg"
ve <- addImg(ve,
  sample_id = "section1",
  image_id = "pomeranian",
  imageSource = url,
  scaleFactor = NA_real_,
  load = TRUE)

grb <- imgGrob(ve,
  sample_id = "section1",
  image_id = "pomeranian")
grid::grid.draw(grb)


##### construct 'SpatialImage'

dir <- file.path("extdata", "10xVisium", "section1", "spatial")
fnm <- file.path(dir, "tissue_lowres_image.png")
imgPath <- system.file(fnm, package = "SpatialExperiment")
(img <- SpatialImage(path = imgPath))

# load image as a 'grob'
img <- loadImg(img)

# accessors
imgGrob(img)
imgPath(img)
imgUrl(img)

# draw the image
grb <- imgGrob(img)
grid::grid.draw(grb)

# unload it, i.e. the 'grob'
img <- unloadImg(img)
imgGrob(img)
  
}
\author{
Helena L. Crowell
}
