% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialCoords<-,SpatialExperiment-method}
\alias{spatialCoords<-,SpatialExperiment-method}
\alias{spatialCoords<-}
\title{#' @rdname SpatialExperiment-methods
#' @export
setReplaceMethod("scaleFactors", 
                 c("SpatialExperiment", "list"),
                 function(x, sample_id=TRUE, image_id=TRUE, 
                          scaleFactors=as.list(rep(1,1)))
                 {
                     idx <- .get_img_idx(x, sample_id, image_id)
                     imgData(x)$scaleFactor[idx] <- scaleFactors
                 }
)
spatialCoords-setter}
\usage{
\S4method{spatialCoords}{SpatialExperiment}(x) <- value
}
\arguments{
\item{se}{a SpatialExperiment class object}

\item{coords}{a DataFrame with the new spatial coordinates to set.}

\item{sample_id}{}
}
\value{
none
}
\description{
a setter method which sets/replaces the spatial coordinate in a
SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
fakeFishCoords <- cbind(fishCoordinates[,c(1:3)], fishCoordinates[,3])
        colnames(fakeFishCoords) <- c("MyCell_ID", "Irrelevant", "x", "y")
spatialCoords(se) <- fakeFishCoords
spatialCoords(se)
}
