% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{spatialCoords<-,SpatialExperiment-method}
\alias{spatialCoords<-,SpatialExperiment-method}
\title{spatialCoords-setter}
\usage{
\S4method{spatialCoords}{SpatialExperiment}(x) <- value
}
\arguments{
\item{x}{a SpatialExperiment class object}

\item{value}{a DataFrame with the new spatial coordinates to set.}
}
\value{

}
\description{
a setter method which sets/replaces the spatial coordinate in a 
SpatialExperiment class object.
}
