% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialCoords<-,SpatialExperiment-method}
\alias{spatialCoords<-,SpatialExperiment-method}
\alias{spatialCoords<-}
\title{spatialCoords-setter}
\usage{
\S4method{spatialCoords}{SpatialExperiment}(se, coords = DataFrame(), sample_id = TRUE) <- value
}
\arguments{
\item{se}{a SpatialExperiment class object}

\item{coords}{a DataFrame with the new spatial coordinates to set.}

\item{sample_id}{}
}
\value{
none
}
\description{
a setter method which sets/replaces the spatial coordinate in a
SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
fakeFishCoords <- cbind(fishCoordinates[,c(1:3)], fishCoordinates[,3])
        colnames(fakeFishCoords) <- c("MyCell_ID", "Irrelevant", "x", "y")
spatialCoords(se) <- fakeFishCoords
spatialCoords(se)
}
