% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{spatialCoords<-,SpatialExperiment-method}
\alias{spatialCoords<-,SpatialExperiment-method}
\alias{spatialCoords<-}
\title{spatialCoords-setter}
\usage{
\S4method{spatialCoords}{SpatialExperiment}(x) <- value
}
\arguments{
\item{x}{a SpatialExperiment class object}

\item{value}{a DataFrame with the new spatial coordinates to set.}
}
\value{
none
}
\description{
a setter method which sets/replaces the spatial coordinate in a
SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
fakeFishCoords <- cbind(fishCoordinates[,c(1:3)], fishCoordinates[,3])
        colnames(fakeFishCoords) <- c("MyCell_ID", "Irrelevant", "x", "y")
spatialCoords(se) <- fakeFishCoords
spatialCoords(se)
}
