% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisiumExperimentMethods.R
\name{spatialCoords<-,VisiumExperiment-method}
\alias{spatialCoords<-,VisiumExperiment-method}
\title{spatialCoords-setter}
\usage{
\S4method{spatialCoords}{VisiumExperiment}(x) <- value
}
\arguments{
\item{x}{a VisiumExperiment class object.}

\item{value}{a DataFrame of spatial coordinates}
}
\description{
sets the spatial coordinate to a VisiumExperiment class object.
}
\examples{
ve <- readRDS(file=system.file(file.path("extdata", "10x_visium",
                         "ve.RDS"), package="SpatialExperiment"))
newSpCord <- spatialCoords(ve)
newSpCord$in_tissue = 0
spatialCoords(ve) <- newSpCord
}
