% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{spatialCoords,SpatialExperiment-method}
\alias{spatialCoords,SpatialExperiment-method}
\title{spatialCoords-getter}
\usage{
\S4method{spatialCoords}{SpatialExperiment}(x)
}
\arguments{
\item{x}{A SpatialExperiment class object.}
}
\value{
a DataFrame within the spatial coordinates.
}
\description{
a getter method which returns the spatial coordinates previously
stored in a SpatialExperiment class object.
}
\examples{
TBD
}
