% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialCoords,SpatialExperiment-method}
\alias{spatialCoords,SpatialExperiment-method}
\alias{spatialCoords}
\title{spatialCoords-getter}
\usage{
\S4method{spatialCoords}{SpatialExperiment}(se, sample_id = TRUE)
}
\arguments{
\item{se}{A SpatialExperiment class object.}

\item{sample_id}{}
}
\value{
a DataFrame within the spatial coordinates.
}
\description{
a getter method which returns the spatial coordinates previously
stored in a SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
spatialCoords(se)
}
