% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperimentMethods.R
\name{spatialCoordsNames,SpatialExperiment-method}
\alias{spatialCoordsNames,SpatialExperiment-method}
\alias{spatialCoordsNames}
\title{spatialCoordsNames-getter}
\usage{
\S4method{spatialCoordsNames}{SpatialExperiment}(x)
}
\arguments{
\item{x}{a SpatialExperiment class object.}
}
\value{
a vector with the colnames of the spatial coordinates.
}
\description{
getter method for the spatial coordinates names in a
SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
spatialCoordsNames(se)
}
