% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R
\name{spatialData,SpatialExperiment-method}
\alias{spatialData,SpatialExperiment-method}
\title{spatialData}
\usage{
\S4method{spatialData}{SpatialExperiment}(se, colDataCols = NULL, sample_id = TRUE, as.df = FALSE)
}
\arguments{
\item{se}{A SpatialExperiment class object.}

\item{colDataCols}{a character vector indicating additional columns to return 
that can be retrieved from the \code{colData} structure.}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying sample 
identifier(s); here, \code{TRUE} is equivalent to all samples.}

\item{as.df}{logical indicating if the returned structure has to be a 
data.frame (default is FALSE).
and \code{NULL} specifies the first available entry (see details)}
}
\value{
a DataFrame within the spatial coordinates.
}
\description{
a getter method which returns the spatial coordinates 
data structure previously stored in a SpatialExperiment class object.
}
\examples{
example(SpatialExperiment)
spatialData(se)
}
