% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\docType{methods}
\name{SparialExperiment-assays}
\alias{SparialExperiment-assays}
\alias{molecules}
\alias{molecules<-}
\alias{molecules,SingleCellExperiment-method}
\alias{molecules<-,SingleCellExperiment-method}
\title{Named assay getters and setters}
\description{
These are methods for getting or setting \code{assay(se, i=X, ...)} 
where \code{se} is a \linkS4class{SpatialExperiment} object and \code{X} is the name of the method.
For example, \code{molecules} will get or set \code{X="molecules"}.
This provides some convenience for users as well as encouraging standardization of assay names across packages.
}
\section{}{
NA
}

\examples{

}
\seealso{
\code{\link{assay}} and \code{\link{assay<-}}, for the wrapped methods.
}
\author{
Dario Righelli
}
