% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\name{SpatialExperiment-colData}
\alias{SpatialExperiment-colData}
\alias{colData}
\alias{colData<-}
\alias{colData<-,SpatialExperiment,DataFrame-method}
\alias{colData<-,SpatialExperiment,NULL-method}
\title{SpatialExperiment colData}
\usage{
\S4method{colData}{SpatialExperiment,DataFrame}(x) <- value

\S4method{colData}{SpatialExperiment,`NULL`}(x) <- value
}
\arguments{
\item{x}{a \code{\link{SpatialExperiment}}}

\item{value}{a \code{\link[S4Vectors]{DataFrame}}}
}
\value{
a SpatialExperiment object
}
\description{
The colData setter expects a DataFrame with a \code{sample_id} column reflecting the
already existing \code{sample_id}(s) present in the object.
An additional check is made on the \code{imgData} data structure.
In case a \code{NULL} value is passed, the \code{colData} are dropped.
}
\examples{
example(SpatialExperiment)
colData(se) 
colData(se) <- NULL
colData(se) 
}
