% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-combine.R
\docType{methods}
\name{SpatialExperiment-combine}
\alias{SpatialExperiment-combine}
\alias{cbind,SingleCellExperiment-method}
\alias{cbind,SpatialExperiment-method}
\title{Combining SpatialExperiment objects}
\usage{
\S4method{cbind}{SpatialExperiment}(..., deparse.level = 1)
}
\value{
a combined SpatialExperiment object
}
\description{
An overview of methods to combine multiple \linkS4class{SpatialExperiment} 
objects by column.
This method is useful for ensuring that all data fields remain 
synchronized multiple SpatialExperiment objects are combined together, 
such as in case of multiple samples.
}
\section{Combining}{

The \code{...} argument is thought to contain one or more 
\linkS4class{SpatialExperiment} objects.
\describe{
\item{\code{cbind(..., deparse.level=1)}:}{
Returns a SpatialExperiment where
all objects in \code{...} are combined column-wise,
i.e., columns in successive objects are appended to the first object.

Each \code{SpatialExperiment} object in \code{...} must have the same 
\code{colData} (with same \code{spatialCoordinates}).
In case of equal sample_id the method will proceed them providing new 
unique ones.
Additionally the \code{imgData} across all the object are \code{rbind}ed. 

Refer to \code{?"\link{cbind,SingleCellExperiment-method}"} for details on 
how metadata and others inherited attributes are combined in the output 
object.
Refer to \code{?\link[base]{cbind}} for the interpretation of \code{deparse.level}.
}
}
}

\examples{
example(SpatialExperiment, echo=FALSE) # using the class example

# Combining:
se1 <- se
se1$sample_id <- "foo1"
cbind(se, se1)

}
\author{
Dario Righelli
}
