% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R, R/imgData.R
\name{colData<-,SpatialExperiment,DataFrame-method}
\alias{colData<-,SpatialExperiment,DataFrame-method}
\alias{imgData}
\alias{imgData<-}
\alias{colData<-}
\alias{scaleFactors}
\alias{spatialCoordsMtx}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{spatialCoordsNames}
\alias{isInTissue}
\alias{colData<-,SpatialExperiment,NULL-method}
\alias{[,SpatialExperiment,ANY,ANY,ANY-method}
\alias{spatialCoords<-,SpatialExperiment-method}
\alias{scaleFactors,SpatialExperiment-method}
\alias{spatialCoords,SpatialExperiment-method}
\alias{spatialCoordsMtx,SpatialExperiment-method}
\alias{spatialCoordsNames,SpatialExperiment-method}
\alias{isInTissue,SpatialExperiment-method}
\alias{imgData,SpatialExperiment-method}
\alias{imgData<-,SpatialExperiment,DataFrame-method}
\alias{imgData<-,SpatialExperiment,NULL-method}
\title{SpatialExperiment methods}
\usage{
\S4method{colData}{SpatialExperiment,DataFrame}(x) <- value

\S4method{colData}{SpatialExperiment,`NULL`}(x) <- value

\S4method{[}{SpatialExperiment,ANY,ANY,ANY}(x, i, j)

\S4method{spatialCoords}{SpatialExperiment}(x) <- value

\S4method{scaleFactors}{SpatialExperiment}(x, sample_id = TRUE, image_id = TRUE)

\S4method{spatialCoords}{SpatialExperiment}(se, sample_id = TRUE)

\S4method{spatialCoordsMtx}{SpatialExperiment}(se, sample_id = TRUE)

\S4method{spatialCoordsNames}{SpatialExperiment}(x)

\S4method{isInTissue}{SpatialExperiment}(x, sample_id = TRUE)

\S4method{imgData}{SpatialExperiment}(x)

\S4method{imgData}{SpatialExperiment,DataFrame}(x) <- value

\S4method{imgData}{SpatialExperiment,`NULL`}(x) <- value
}
\arguments{
\item{x}{a VisiumExperiment class object.}

\item{value}{a \code{\link[S4Vectors]{DataFrame}}}

\item{i, j}{indices specifying elements to extract or replace}

\item{sample_id, }{image_id 
character string, \code{TRUE} or \code{NULL} specifying sample/image 
identifier(s); here, \code{TRUE} is equivalent to all samples/images 
and \code{NULL} specifies the first available entry (see details)}

\item{se}{A SpatialExperiment class object.}

\item{coords}{a DataFrame with the new spatial coordinates to set.}

\item{sample_id}{character string, \code{TRUE} or \code{NULL} specifying sample 
identifier(s); here, \code{TRUE} is equivalent to all samples 
and \code{NULL} specifies the first available entry (see details)}
}
\value{
none

a DataFrame within the spatial coordinates.

a matrix object within the spatial coordinates.

a vector with the colnames of the spatial coordinates.

a TRUE/FALSE mask.
}
\description{
a setter method which sets/replaces the spatial coordinate in a
SpatialExperiment class object.

a getter method which returns the spatial coordinates previously
stored in a SpatialExperiment class object.

a getter method which returns the spatial coordinates previously
stored in a SpatialExperiment class object.

getter method for the spatial coordinates names in a
SpatialExperiment class object.

returns a mask of TRUE/FALSE Barcodes spots, indicating which
ones are in tissue and which ones are not.
}
\examples{
example(SpatialExperiment)
fakeFishCoords <- cbind(fishCoordinates[,c(1:3)], fishCoordinates[,3])
        colnames(fakeFishCoords) <- c("MyCell_ID", "Irrelevant", "x", "y")
spatialCoords(se) <- fakeFishCoords
spatialCoords(se)
example(SpatialExperiment)
spatialCoords(se)
example(SpatialExperiment)
spatialCoordsMtx(se)
example(SpatialExperiment)
spatialCoordsNames(se)
ve <- readRDS(file=system.file(file.path("extdata", "10x_visium",
                         "ve.RDS"), package="SpatialExperiment"))
isInTissue(ve)
sum(isInTissue(ve))
}
