% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R, R/imgData.R
\name{SpatialExperiment-methods}
\alias{SpatialExperiment-methods}
\alias{spatialData}
\alias{spatialData<-}
\alias{spatialDataNames}
\alias{spatialDataNames<-}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{spatialCoordsNames}
\alias{spatialCoordsNames<-}
\alias{imgData}
\alias{imgData<-}
\alias{scaleFactors}
\alias{spatialData,SpatialExperiment-method}
\alias{spatialData<-,SpatialExperiment,DFrame-method}
\alias{spatialData<-,SpatialExperiment,NULL-method}
\alias{spatialDataNames,SpatialExperiment-method}
\alias{spatialDataNames<-,SpatialExperiment,character-method}
\alias{spatialDataNames<-,SpatialExperiment,NULL-method}
\alias{spatialCoords,SpatialExperiment-method}
\alias{spatialCoords<-,SpatialExperiment,matrix-method}
\alias{spatialCoords<-,SpatialExperiment,NULL-method}
\alias{spatialCoordsNames,SpatialExperiment-method}
\alias{spatialCoordsNames<-,SpatialExperiment,character-method}
\alias{spatialCoordsNames<-,SpatialExperiment,NULL-method}
\alias{scaleFactors,SpatialExperiment-method}
\alias{imgData,SpatialExperiment-method}
\alias{imgData<-,SpatialExperiment,DataFrame-method}
\alias{imgData<-,SpatialExperiment,NULL-method}
\title{Methods for spatial attributes}
\usage{
\S4method{spatialData}{SpatialExperiment}(x, spatialCoords = FALSE, colData = FALSE)

\S4method{spatialData}{SpatialExperiment,DFrame}(x) <- value

\S4method{spatialData}{SpatialExperiment,`NULL`}(x) <- value

\S4method{spatialDataNames}{SpatialExperiment}(x)

\S4method{spatialDataNames}{SpatialExperiment,character}(x) <- value

\S4method{spatialDataNames}{SpatialExperiment,`NULL`}(x) <- value

\S4method{spatialCoords}{SpatialExperiment}(x)

\S4method{spatialCoords}{SpatialExperiment,matrix}(x) <- value

\S4method{spatialCoords}{SpatialExperiment,`NULL`}(x) <- value

\S4method{spatialCoordsNames}{SpatialExperiment}(x)

\S4method{spatialCoordsNames}{SpatialExperiment,character}(x) <- value

\S4method{spatialCoordsNames}{SpatialExperiment,`NULL`}(x) <- value

\S4method{scaleFactors}{SpatialExperiment}(x, sample_id = TRUE, image_id = TRUE)

\S4method{imgData}{SpatialExperiment}(x)

\S4method{imgData}{SpatialExperiment,DataFrame}(x) <- value

\S4method{imgData}{SpatialExperiment,`NULL`}(x) <- value
}
\arguments{
\item{x}{A \code{\link{SpatialExperiment}} object.}

\item{spatialCoords}{Logical specifying whether to include columns from
\code{spatialCoords} in the output \code{\link{DataFrame}} from
\code{spatialData}. Default = \code{FALSE}.}

\item{colData}{Logical or character vector specifying any additional columns
from \code{colData} to include in the output \code{\link{DataFrame}} from
\code{spatialData}. Default = \code{FALSE} (no columns).}

\item{value}{Replacement value for replacement methods.}

\item{sample_id}{Logical value or character vector specifying sample
identifier(s) for \code{scaleFactors}. Default = \code{TRUE} (all samples).}

\item{image_id}{Logical value or character vector specifying image
identifier(s) for \code{scaleFactors}. Default = \code{TRUE} (all images).}
}
\value{
see methods details.
}
\description{
The \code{\link{SpatialExperiment}} class provides a family of methods to get
and set spatial data attributes in \code{\link{SpatialExperiment}} objects.
Spatial attributes include \code{spatialData}, \code{spatialCoords},
\code{imgData}, and \code{scaleFactors}.
}
\details{
Additional details for each type of data attribute are provided below.



We distinguish the spatial data from the spatial coordinates as described in 
this paragraph.
With spatial data we refer to the general data structure containing the 
spatial coordinates, while we refer for spatial coordinates only to the 
specific spatial coordinates columns (as x, y, z, ...).
}
\section{spatialData Methods}{

All the methods described in this section takes as input the \code{sample_id}
argument.
See the parameter description for additional information.
\describe{
\item{\code{spatialData(x) <- value}:}{ The spatialData setter expects a 
data.frame/DataFrame object named with the
previously defined colnames for the spatial coordinates.
It is possible to get/set the spatial coordinates names with the 
spatialCoordsNames getter/setter, by default they are set by the 
SpatialExperiment constructor to c("x", "y").
If the input spatial data structure doesn't contain an in_tissue column, 
a default in_tissue column is set to 1.
If NULL is passed as value the all the related data structures are set 
accordingly.}
\item{\code{spatialData(x)}:}{ The spatialData getter gives the possibility 
to retrieve also additional columns from the
colData structure by the aid of the \code{cd_bind} argument.
It returns by default a DataFrame, but it is possible to retrieve a 
matrix by setting \code{as_df=FALSE}.}
\item{\code{spatialCoords(x)}:}{ The \code{spatialCoords} getter returns the 
spatial coordinates in a matrix-like form, but it is possible to retrieve 
them in a DataFrame form by setting \code{as_df=TRUE}.}
}
}

\section{spatialNames Methods}{

\describe{
\item{\code{spatialDataNames(x)}:}{ Returns the names of the spatialData 
structure.}
\item{\code{spatialCoordsNames(x)<-value}:}{ Sets the names of the spatial 
coordinates accordingly with the assigned value.}
\item{\code{spatialCoordsNames(x)}:}{ Returns the stored names of the spatial 
coordinates.}
}
}

\section{imgData Methods}{

\describe{
\item{\code{imgData(x)<-value}:}{ This setter allows to set a DataFrame 
object as imgData of the SpatialExperiment object class.}
\item{\code{imgData(x)}:}{ This getter returns the imgData DataFrame.}
}
}

\section{Other Methods}{

\describe{
\item{\code{scaleFactors(x, sample_id, image_id)}:}{ This getter returns 
the scale factors associated to the \code{sample_id}(s) 
and \code{image_id}(s) passed as input.
This is related to the stored image(s) into the SpatialExperiment.
\code{imgData} structure.
See the \code{sample_id} and \code{image_id} section description for further 
details.}
}
}

\examples{
example(SpatialExperiment)
## The spatialData methods
fakeSpData <- spatialData(spe)
fakeSpData$array_col <- fakeSpData$array_row
spatialData(spe) <- fakeSpData

# spatialData returns a DataFrame
head(spatialData(spe))

# spatialCoords returns a matrix of coordinates
head(spatialCoords(spe))

# combine spatialData with spatialCoords & colData of interest
head(spatialData(spe, spatialCoords=TRUE, colData="sample_id"))

# changing spatial coordinates names
spatialCoordsNames(spe)
spatialCoordsNames(spe) <- c("x","y")
spatialCoordsNames(spe)
head(spatialCoords(spe))

# imgData
imgData(spe)
scaleFactors(spe)

# tabulate number of spots mapped to tissue
table(
  in_tissue = spe$in_tissue, 
  sample_id = spe$sample_id)

}
