% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R, R/imgData.R
\name{colData<-,SpatialExperiment,NULL-method}
\alias{colData<-,SpatialExperiment,NULL-method}
\alias{imgData}
\alias{imgData<-}
\alias{colData<-}
\alias{scaleFactors}
\alias{[,SpatialExperiment,ANY,ANY,ANY-method}
\alias{scaleFactors,SpatialExperiment-method}
\alias{imgData,SpatialExperiment-method}
\alias{imgData<-,SpatialExperiment,DataFrame-method}
\alias{imgData<-,SpatialExperiment,NULL-method}
\title{SpatialExperiment methods}
\usage{
\S4method{colData}{SpatialExperiment,`NULL`}(x) <- value

\S4method{[}{SpatialExperiment,ANY,ANY,ANY}(x, i, j)

\S4method{scaleFactors}{SpatialExperiment}(x, sample_id = TRUE, image_id = TRUE)

\S4method{imgData}{SpatialExperiment}(x)

\S4method{imgData}{SpatialExperiment,DataFrame}(x) <- value

\S4method{imgData}{SpatialExperiment,`NULL`}(x) <- value
}
\arguments{
\item{x}{a \code{\link{SpatialExperiment}}}

\item{value}{a \code{\link[S4Vectors]{DataFrame}}}

\item{i, j}{indices specifying elements to extract or replace}

\item{sample_id, }{image_id 
character string, \code{TRUE} or \code{NULL} specifying sample/image 
identifier(s); here, \code{TRUE} is equivalent to all samples/images 
and \code{NULL} specifies the first available entry (see details)}
}
\description{
SpatialExperiment methods
}
