% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment-methods.R, R/imgData.R
\name{SpatialExperiment-methods}
\alias{SpatialExperiment-methods}
\alias{spatialData}
\alias{spatialData<-}
\alias{spatialDataNames}
\alias{spatialDataNames<-}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{spatialCoordsNames}
\alias{spatialCoordsNames<-}
\alias{imgData}
\alias{imgData<-}
\alias{scaleFactors}
\alias{spatialData,SpatialExperiment-method}
\alias{spatialData<-,SpatialExperiment,DFrame-method}
\alias{spatialData<-,SpatialExperiment,NULL-method}
\alias{spatialDataNames,SpatialExperiment-method}
\alias{spatialDataNames<-,SpatialExperiment,character-method}
\alias{spatialCoords,SpatialExperiment-method}
\alias{spatialCoords<-,SpatialExperiment,matrix-method}
\alias{spatialCoords<-,SpatialExperiment,NULL-method}
\alias{spatialCoordsNames,SpatialExperiment-method}
\alias{spatialCoordsNames<-,SpatialExperiment,character-method}
\alias{spatialCoordsNames<-,SpatialExperiment,NULL-method}
\alias{scaleFactors,SpatialExperiment-method}
\alias{imgData,SpatialExperiment-method}
\alias{imgData<-,SpatialExperiment,DataFrame-method}
\alias{imgData<-,SpatialExperiment,NULL-method}
\title{Methods for spatial attributes}
\usage{
\S4method{spatialData}{SpatialExperiment}(x, spatialCoords = FALSE)

\S4method{spatialData}{SpatialExperiment,DFrame}(x) <- value

\S4method{spatialData}{SpatialExperiment,`NULL`}(x) <- value

\S4method{spatialDataNames}{SpatialExperiment}(x)

\S4method{spatialDataNames}{SpatialExperiment,character}(x) <- value

\S4method{spatialCoords}{SpatialExperiment}(x)

\S4method{spatialCoords}{SpatialExperiment,matrix}(x) <- value

\S4method{spatialCoords}{SpatialExperiment,`NULL`}(x) <- value

\S4method{spatialCoordsNames}{SpatialExperiment}(x)

\S4method{spatialCoordsNames}{SpatialExperiment,character}(x) <- value

\S4method{spatialCoordsNames}{SpatialExperiment,`NULL`}(x) <- value

\S4method{scaleFactors}{SpatialExperiment}(x, sample_id = TRUE, image_id = TRUE)

\S4method{imgData}{SpatialExperiment}(x)

\S4method{imgData}{SpatialExperiment,DataFrame}(x) <- value

\S4method{imgData}{SpatialExperiment,`NULL`}(x) <- value
}
\arguments{
\item{x}{A \code{\link{SpatialExperiment}} object.}

\item{spatialCoords}{Logical specifying whether to include columns from
\code{spatialCoords} in the output \code{\link{DataFrame}} from
\code{spatialData}. Default = \code{FALSE}.}

\item{value}{Replacement value for replacement methods.}

\item{sample_id}{Logical value or character vector specifying sample
identifier(s) for \code{scaleFactors}. Default = \code{TRUE} (all samples).}

\item{image_id}{Logical value or character vector specifying image
identifier(s) for \code{scaleFactors}. Default = \code{TRUE} (all images).}

\item{spatialData}{Logical specifying whether to include columns from
\code{spatialData} in the output \code{\link{DataFrame}} from
\code{colData}. Default = \code{FALSE}}
}
\value{
Return value varies depending on method, as described below.
}
\description{
The \code{\link{SpatialExperiment}} class provides a family of methods to get
and set spatial data attributes in \code{\link{SpatialExperiment}} objects.
Spatial attributes include \code{spatialData}, \code{spatialCoords},
\code{imgData}, and \code{scaleFactors}.
}
\details{
Additional details for each type of data attribute are provided below.

\code{\link{spatialData}} and \code{\link{spatialCoords}} are distinguished
as follows: \code{spatialData} is a \code{DataFrame} containing all the data
associated with the spatial information (optionally including spatial
coordinates from \code{spatialCoords}), while \code{spatialCoords} is a
numeric matrix containing only the defined spatial coordinates (e.g. columns
\code{x} and \code{y}).
}
\section{spatialData and spatialCoords methods}{

\describe{
\item{\code{spatialData(x)}: }{The \code{spatialData} getter provides the 
optional argument \code{spatialCoords}, which can be used to include the 
columns of spatial coordinates (\code{spatialCoords}) in the output 
\code{DataFrame}.}
\item{\code{spatialData(x) <- value}: }{The \code{spatialData} setter
expects a \code{DataFrame}. If the input does not contain an \code{in_tissue} 
column, this will be included with a default value of \code{1}.}
\item{\code{spatialCoords(x)}: }{Getter for numeric matrix of spatial
coordinates.}
\item{\code{spatialCoords(x) <- value}: }{Setter for numeric matrix of
spatial coordinates.}
}
}

\section{spatialDataNames and spatialCoordsNames methods}{

\describe{
\item{\code{spatialDataNames(x)}: }{Returns the column names of the
\code{spatialData} \code{DataFrame}.}
\item{\code{spatialDataNames(x) <- value}: }{Setter to replace column names
in the \code{spatialData} \code{DataFrame}.}
\item{\code{spatialCoordsNames(x)}: }{Returns the defined names of the
spatial coordinates (e.g. \code{c("x", "y")}).}
\item{\code{spatialCoordsNames(x) <- value}: }{Setter to define the names of
the spatial coordinate columns.}
}
}

\section{imgData methods}{

\describe{
\item{\code{imgData(x)}: }{Getter to return the \code{imgData} \code{DataFrame}.}
\item{\code{imgData(x) <- value}: }{Setter to provide a \code{DataFrame}
object as \code{imgData} of the \code{SpatialExperiment} object.}
}
}

\section{Other methods}{

\describe{
\item{\code{scaleFactors(x, sample_id, image_id)}: }{Getter to return the
scale factors associated with the \code{sample_id}(s) and \code{image_id}(s)
provided. This is related to the stored image(s) in the
\code{SpatialExperiment} \code{imgData} structure. See argument descriptions
for further details.}
}
}

\examples{
example(read10xVisium)

# spatialData returns a DataFrame
spatialData(spe)

# spatialCoords returns a numeric matrix
head(spatialCoords(spe))

# spatialData replacement method
spdata <- spatialData(spe)
spdata$array_col <- spdata$array_row
spatialData(spe) <- spdata

# return additional columns for spatialData
spatialData(spe, spatialCoords=TRUE)

# change spatial coordinate names
spatialCoordsNames(spe)
spatialCoordsNames(spe) <- c("x", "y")
head(spatialCoords(spe))

# imgData and scale factors
imgData(spe)
scaleFactors(spe)

# tabulate number of spots mapped to tissue
cd <- colData(spe, spatialData = TRUE)
table(
  in_tissue = cd$in_tissue, 
  sample_id = cd$sample_id)

}
