% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{SpatialExperiment-misc}
\alias{SpatialExperiment-misc}
\alias{show,SpatialExperiment-method}
\title{Miscellaneous SpatialExperiment methods}
\usage{
\S4method{show}{SpatialExperiment}(object)
}
\arguments{
\item{object}{a \code{\link{SpatialExperiment}} object}
}
\value{
Returns NULL
}
\description{
Miscellaneous methods for the \code{\link{SpatialExperiment}} class and its
descendants that do not fit into any other documentation category such as,
for example, show methods.
}
\examples{
example(read10xVisium)
spe
}
\author{
Dario Righelli and Helena L. Crowell
}
