% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{SpatialExperiment-subset}
\alias{SpatialExperiment-subset}
\alias{[,SpatialExperiment,ANY,ANY,ANY-method}
\title{SpatialExperiment subset}
\usage{
\S4method{[}{SpatialExperiment,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{a SpatialExperiment object}

\item{i}{the row indexes to subset}

\item{j}{the col indexes to subset}
}
\description{
The subset method ensures that all the subsetted rows/cols 
are updated across the colData and spatialData objects
}
\section{subset}{

\describe{
\itemize{
\item{\code{[}:} subsetting method
}
}
}

\examples{
example(SpatialExperiment)
se1 <- se[,1:10]
colData(se1)
spatialData(se1)
}
