% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SpatialExperiment-class}
\alias{SpatialExperiment-class}
\alias{SpatialExperiment}
\title{The SpatialExperiment class}
\usage{
SpatialExperiment(..., spatialCoords = data.frame())
}
\arguments{
\item{...}{arguments to be passed to the \code{\link{SingleCellExperiment}} 
constructor to fill the slots of the base class.}

\item{spatialCoords}{the 10x Visium spatial coordinates}
}
\description{
The SpatialExperiment class is designed to represent 10x Visium spatial 
Gene Expression data.
It inherits from the \linkS4class{SingleCellExperiment} class and is used in 
the same manner.
In addition, the class supports the integration with 10x Visium spatial 
coordinates and its scale factors.
}
\section{Slots}{

\describe{
\item{\code{int_spcIdx}}{integer.}
}}

\examples{
TBD
}
\author{
Dario Righelli
}
