% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SpatialExperiment.R
\docType{class}
\name{SpatialExperiment-class}
\alias{SpatialExperiment-class}
\alias{SpatialExperiment}
\title{The SpatialExperiment class}
\usage{
SpatialExperiment(
  ...,
  sample_id = "Sample01",
  spatialCoords = DataFrame(),
  scaleFactors = 1,
  imageSources = NULL,
  image_id = NULL,
  loadImage = TRUE,
  imgData = NULL
)
}
\arguments{
\item{...}{arguments to be passed to the \code{\link{SingleCellExperiment}} 
constructor to fill the slots of the base class.}

\item{sample_id}{a character vector of unique sample identifiers in
conformity with the \code{sample_id}s in \code{imgData}}

\item{imgData}{a \code{DataFrame} storing the image data (see details)}

\item{xyzData}{a two/(three)-column numeric matrix containing 
spatial xy(z)-coordinates for each observation}

\item{inTissue}{a logical vector indicating whether or not
an observation could be mapped onto the tissue}
}
\value{
a \code{SpatialExperiment}
}
\description{
The SpatialExperiment class is designed to represent spatial transcriptomics 
(ST) data. It inherits from the \linkS4class{SingleCellExperiment} class and
is used in the same manner. In addition, the class supports storage of images
for multiple samples and of different resolutions via \code{\link{imgData}}, 
and requires certain observation metadata specific to ST data to be present.
}
\details{
If any of arguments \code{sample_id}, \code{inTissue} and \code{xyzData}
are missing, the \code{SpatialExperiment} constructor will assume 
these are supplied via the input \code{colData}.
}
\examples{
dir <- system.file(
  file.path("extdata", "10xVisium", "section1"),
  package = "SpatialExperiment")

# read in counts
fnm <- file.path(dir, "raw_feature_bc_matrix.h5")
sce <- DropletUtils::read10xCounts(fnm)

# read in image data
img <- readImgData(
  path = file.path(dir, "spatial"), 
  sample_id="foo")

# read in spatial coordinates
fnm <- file.path(dir, "spatial", "tissue_positions_list.csv")
xyz <- read.csv(fnm, header = FALSE, row.names = 1, 
  col.names = c(
    "barcode", "inTissue", "array_row", "array_col", 
    "pxl_row_in_fullres", "pxl_col_in_fullres"))
   
# construct observation & feature metadata 
rd <- S4Vectors::DataFrame(
  symbol = rowData(sce)$Symbol) 
  
cd <- S4Vectors::DataFrame(
  sample_id = "foo",
  inTissue = as.logical(xyz$inTissue),
  xyzData = I(as.matrix(xyz[, grep("pxl", names(xyz))])))
  
# construct 'SpatialExperiment'
SpatialExperiment(
  assays = list(counts = assay(sce)),
  colData = cd, rowData = rd, imgData = img)
}
\author{
Dario Righelli & Helena L. Crowell
}
