% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{SpatialImage-misc}
\alias{SpatialImage-misc}
\alias{.spi_show}
\alias{show,SpatialImage-method}
\title{Miscellaneous \code{SpatialImage} methods}
\usage{
.spi_show(object)

\S4method{show}{SpatialImage}(object)
}
\arguments{
\item{object}{a \code{SpatialImage}.}
}
\value{
none.
}
\description{
Miscellaneous methods for the \code{\link{SpatialImage}} class that do not 
fit in any other documentation category such as, for example, show methods.
}
\author{
Helena L. Crowell
}
