% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{SpatialImage-misc}
\alias{SpatialImage-misc}
\alias{show,VirtualSpatialImage-method}
\title{Miscellaneous \code{SpatialImage} methods}
\usage{
\S4method{show}{VirtualSpatialImage}(object)
}
\arguments{
\item{object}{a \code{SpatialImage} object}
}
\value{
none
}
\description{
Miscellaneous methods for the \code{\link{SpatialImage}} class that do not
fit into any other documentation category such as, for example, show methods.
}
\author{
Helena L. Crowell
}
